토큰나이즈 함 ㅗ, 겹괄호랑 따옴표처리 해야함 ㅗ^^ㅗ 미좋 ^^ㅗㅗㅗㅗㅗㅗㅗ 하하하ㅏㅎ하하ㅏㅎ

+++ libft/ft_strndup.c
@@ -1,6 +1,6 @@
 #include "libft.h"
 
-char	*ft_strndup(const char *s1, int n)
+char	*ft_strndup(const char *s1, size_t n)
 {
 	char	*dup;
 	size_t	s1len;
